moved usIsRegisteredForDay to own file
